Refactor file paths for BaseTool and update imports in various feature files
