+++ js/features/password-generator.js
@@ -1,7 +1,6 @@
-import { BaseTool } from './base-tool.js';
-import { notifications } from '../utils/ui.js';
+import { BaseTool } from '../utils/base-tool.js';
+import { showNotification } from '../utils/ui.js';
 import { STORAGE_KEYS } from '../utils/constants.js';
-import utils from '../utils/helpers.js';
 
 class PasswordGenerator extends BaseTool {
     constructor() {
